Move env_save to cover the everest tests
